chore(packages/text): move typography to packages
