+++ packages/components/stamp/src/stamp.story.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Section from '../../../../.storybook/decorators/section';
-import Text from '../../../../src/components/typography/text';
+import Text from '../../text';
 import Spacings from '../../../../src/components/spacings';
 import * as icons from '../../icons';
 import Stamp, { availableTones } from './stamp';
