+++ packages/components/text/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './text';
+
+export { default as version } from './version';
