+++ packages/components/text/src/text.js
