+++ packages/components/text/src/version.js
