+++ packages/components/text/typography.story.js
@@ -3,8 +3,8 @@ import styled from '@emotion/styled';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, boolean, select, text } from '@storybook/addon-knobs/react';
 import Section from '../../../.storybook/decorators/section';
-import Text from './text';
-import Readme from './text/README.md';
+import Text from './src/text';
+import Readme from './README.md';
 
 const InlineColorWrapper = styled.div`
   background-color: #e1ffdd;
