--- src/components/typography/text/index.js
@@ -1,1 +1,0 @@
-export { default } from './text';
